Move tests to match src path
